% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants.R
\docType{data}
\name{ants}
\alias{ants}
\title{Example dataset}
\format{
A data frame with 29 rows and seven variables

#' @source \url{https://doi.org/10.1111/2041-210X.13568}
}
\usage{
ants
}
\description{
Empirical data: seven morphological variables from 29 ant species
}
\keyword{datasets}
