% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex0.R
\docType{data}
\name{ex0}
\alias{ex0}
\title{Example dataset}
\format{
A data frame with 100 rows and five variables
}
\usage{
ex0
}
\description{
Simulated data: five uncorrelated variables and 100 observations
}
\keyword{datasets}
