% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex025.R
\docType{data}
\name{ex025}
\alias{ex025}
\title{Example dataset}
\format{
A data frame with 100 rows and five variables
}
\usage{
ex025
}
\description{
Simulated data: five correlated variables (r = 0.25) and 100 observations
}
\keyword{datasets}
