% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex05.R
\docType{data}
\name{ex05}
\alias{ex05}
\title{Example dataset}
\format{
A data frame with 100 rows and five variables
}
\usage{
ex05
}
\description{
Simulated data: five correlated variables (r = 0.50) and 100 observations
}
\keyword{datasets}
