% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAtest.R
\name{PCAtest}
\alias{PCAtest}
\title{PCAtest: Statistical significance of PCA}
\usage{
PCAtest(x, nperm = 1000, nboot = 1000, alpha = 0.05, corr = F, plot = T)
}
\arguments{
\item{x}{matrix}

\item{nperm}{numeric}

\item{nboot}{numeric}

\item{alpha}{numeric}

\item{corr}{numeric}

\item{plot}{logical}
}
\description{
PCAtest: Statistical significance of PCA
}
\examples{
v1<-runif(100,0,1)
v2<-runif(100,0,1)
x<-cbind(v1,v2)
PCAtest(x, 1000, 1000, 0.05, corr=FALSE, plot=TRUE)
}
